Changes next piece selection to closer to games
In most tetris games, the selection of the next piece is not purely random.
It follows a simple algorythm that can be likened to drawing pieces from
a bag until it is empty, then re-filling the bag. In effect, this means that
all 7 pieces must be selected before one of them can be selected again.

For example, if 2 "T" pieces are selected in a row, then another "T" won't
be selected for at least 6 other pieces, and it is guaranteed to be selected
after 12 pieces.

We believe this won't impact the agent too much, but opens up some new
possibilities for agent strategies.