Fixes bug in clearing multiple lines
The bug happens when you clear more than one line. The previous code
didn't guarantee that the lines were cleared in the correct order.
The correct order should be from top to bottom, or else an indexation
problem occurs due to dropping blocks.

For example, if we delete lines 29 and 28, in this order, then line 29
will be cleared, line 28 will became 29 (and consequently 27 into 28),
then line 28 will also be cleared (ex-27). In effect, the game deleted the
wrong line and the line that should have been deleted remains in play.

P.S.: Thanks to Dinis, Isabel and Camila (NMECS: 98452, 93343, 97880) for
helping us track and fix this bug.